+++ shuup/admin/static_src/base/js/datetimepicker.js
@@ -9,18 +9,23 @@
 $(function() {
     $(".form-control.datetime").datetimepicker({
         format: window.ShuupAdminConfig.settings.datetimeInputFormat,
-        step: window.ShuupAdminConfig.settings.datetimeInputStep
+        step: window.ShuupAdminConfig.settings.datetimeInputStep,
+        scrollMonth : false, // Scrolling is disabled due to it being broken (always going down)
+        scrollInput : false
     });
 
     $(".form-control.date").datetimepicker({
         format: window.ShuupAdminConfig.settings.dateInputFormat,
-        timepicker: false
+        timepicker: false,
+        scrollMonth : false,
+        scrollInput : false
     });
 
     $(".form-control.time").datetimepicker({
         format: window.ShuupAdminConfig.settings.timeInputFormat,
+        step: window.ShuupAdminConfig.settings.datetimeInputStep,
         datepicker: false,
-        step: window.ShuupAdminConfig.settings.datetimeInputStep
+        scrollInput : false
     });
 
     jQuery.datetimepicker.setLocale(window.ShuupAdminConfig.settings.dateInputLocale);